backend/Models/requirements.txt
# Core Machine Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
timm>=0.9.2
scikit-learn>=1.0.0

# Audio Processing
librosa>=0.9.2
audiomentations>=0.25.0
soundfile>=0.12.1

# Image Processing
opencv-python>=4.7.0
albumentations>=1.3.0
mediapipe>=0.10.0
facenet-pytorch>=2.5.2
pillow>=9.0.0

# Data Handling
numpy>=1.22.0
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.12.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0.0
wandb>=0.15.0
python-dotenv>=1.0.0

# Optional for improved performance
nvidia-ml-py>=11.525.84  # For GPU monitoring
tensorboard>=2.10.0      # Alternative to wandb

# Video Processing
imageio>=2.25.0
imageio-ffmpeg>=0.4.7
moviepy>=1.0.3